Added nulll checker while reading cached timeStamp. Adapting methods to read cache by exact key or pattern key.

+++ src/app/utils.tsx
@@ -1,18 +1,24 @@
 export const isOneDayDiff = (timeStampKey) =>
 {
     const timeStamp = localStorage.getItem(timeStampKey);
+    if (timeStamp === null)
+        return false
+
     const timeDiff = ((new Date()).getTime() - timeStamp) / 1000;
     return timeDiff > 86000;
 }
 
-export const readCachedData = (timeStampKey: string, keyPattern: string) => 
+export const readCachedDataByPattern = (timeStampKey: string, keyPattern: string) => 
 {
-    const allCachedData = {...localStorage};
     var cacheData: [] = [];
 
     if (isOneDayDiff(timeStampKey))
     {
-        for (const [key, value] of Object.entries(allCachedData)) 
+        localStorage.clear();
+    }
+    else
+    {
+        for (const [key, value] of Object.entries(localStorage)) 
         {
             if (key.match( keyPattern))
             {
@@ -21,10 +27,22 @@ export const readCachedData = (timeStampKey: string, keyPattern: string) =>
             }
         }
     }
-    else
+
+    return cacheData;
+}
+
+export const readCachedData = (timeStampKey: string, key: string) =>
+{
+    var cacheData: any = {};
+
+    if (isOneDayDiff(timeStampKey))
     {
         localStorage.clear();
     }
+    else
+    {
+        cacheData = localStorage.getItem(key);
+    }
 
-    return cacheData;
+    return JSON.parse(cacheData);
 }
